move Bugzilla and Jira enabling fixtures to the top level to allow inter-module sharing
